Revert "Jung min generic login"
