+++ src/main/java/com/multi/fourtunes/model/dto/UserDto.java
@@ -22,4 +22,6 @@ public class UserDto {
 	private String user_grade;
 	private int membership_duration;
 	private Date pay_date;
+	
+	
 }
