+++ src/main/resources/templates/index.html
@@ -1,88 +1,78 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-<meta charset="UTF-8" />
-<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-<link rel="stylesheet" href="/resources/css/main.css" />
-<link rel="stylesheet" href="/resources/css/nav.css" />
-<link rel="stylesheet" href="/resources/css/globalfont.css" />
-<link rel="shortcut icon" href="/resources/img/realfavicon.ico" />
-<title>4Tunes</title>
-</head>
-<body>
-	<div id="nav">
-		<div class="banner">
-			<a href="/"><img src="/resources/img/banner.png" /></a>
-		</div>
-		<div id="login">
-			<div>
-				<a
-					th:href="${currentUser.user_id != null ? '/login/logout' : '/nav/login'}">
-					<img class="login_img" src="/resources/img/usertest.png" /> <span
-					th:text="${currentUser.user_id != null ? 'LOGOUT' : 'LOGIN'}"></span>
-				</a>
-			</div>
-			<br />
-			<div>
-				<a href="/nav/mypage"><img class="mypage_img"
-					src="/resources/img/file.png" />MYPAGE</a>
-			</div>
-		</div>
-		<div class="search-container">
-			<span class="search-icon"><img src="/resources/img/search.png"
-				class="search-icon" onclick="location.href='/nav/search'" /></span> <input
-				type="text" class="search-input" placeholder="키워드 입력" />
-		</div>
-		<div>
-			<a href="/nav/suggested">맞춤추천</a>
-		</div>
-		<div>
-			<a href="/nav/chart">인기차트</a>
-		</div>
-		<div>
-			<a href="/nav/community">커뮤니티</a>
-		</div>
-		<div>
-			<a href="/nav/membership">멤버십</a>
-		</div>
-	</div>
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/resources/css/main.css" />
+    <link rel="stylesheet" href="/resources/css/nav.css" />
+    <link rel="stylesheet" href="/resources/css/globalfont.css" />
+    <link rel="shortcut icon" href="/resources/img/realfavicon.ico" />
+    <title>4Tunes</title>
+  </head>
+  <body>
+    <div id="nav">
+      <div class="banner">
+        <a href="/"><img src="/resources/img/banner.png" /></a>
+      </div>
+      <div id="login">
+        <div>
+          <a href="/nav/login"
+            ><img class="login_img" src="/resources/img/usertest.png" />LOGIN</a
+          >
+        </div>
+        <br />
+        <div>
+          <a href="/nav/mypage"
+            ><img class="mypage_img" src="/resources/img/file.png" />MYPAGE</a
+          >
+        </div>
+      </div>
+      <div class="search-container">
+        <span class="search-icon"
+          ><img src="/resources/img/search.png" class="search-icon" onclick="location.href='/nav/search'"
+        /></span>
+        <input type="text" class="search-input" placeholder="키워드 입력" />
+      </div>
+      <div><a href="/nav/suggested">맞춤추천</a></div>
+      <div><a href="/nav/chart">인기차트</a></div>
+      <div><a href="/nav/community">커뮤니티</a></div>
+      <div><a href="/nav/membership">멤버십</a></div>
+    </div>
 
-	<div id="main">
-		<div id="playlist">
-			<h1>플레이리스트</h1>
-			<div class="playlist">
-				<a href="/innerpaging/playlist/mine">내 재생목록<img src="" alt="" /></a>
-			</div>
-			<div class="playlist">
-				<a href="/innerpaging/playlist/public"><img src="" alt="" />public
-					재생목록</a>
-			</div>
-			<div class="playlist">
-				<a href=""><img src="" alt="" /></a>
-			</div>
-			<div class="playlist">
-				<a href=""><img src="" alt="" /></a>
-			</div>
-		</div>
+    <div id="main">
+      <div id="playlist">
+        <h1>플레이리스트</h1>
+        <div class="playlist">
+          <a href="/innerpaging/playlist/mine">내 재생목록<img src="" alt="" /></a>
+        </div>
+        <div class="playlist">
+          <a href="/innerpaging/playlist/public"><img src="" alt="" />public 재생목록</a>
+        </div>
+        <div class="playlist">
+          <a href=""><img src="" alt="" /></a>
+        </div>
+        <div class="playlist">
+          <a href=""><img src="" alt="" /></a>
+        </div>
+      </div>
 
-		<div id="today">
-			<h1>오늘의 선곡</h1>
-			<div class="today">
-				<a href="innerpaging/playlist/suggested">오늘의 선곡<img src=""
-					alt="" /></a>
-			</div>
-			<div class="today">
-				<a href=""><img src="" alt="" /></a>
-			</div>
-			<div class="today"></div>
-			<div class="today">
-				<a href=""><img src="" alt="" /></a>
-			</div>
-			<div class="today">
-				<a href=""><img src="" alt="" /></a>
-			</div>
-		</div>
-	</div>
-</body>
+      <div id="today">
+        <h1>오늘의 선곡</h1>
+        <div class="today">
+          <a href="innerpaging/playlist/suggested">오늘의 선곡<img src="" alt="" /></a>
+        </div>
+        <div class="today">
+          <a href=""><img src="" alt="" /></a>
+        </div>
+        <div class="today"></div>
+        <div class="today">
+          <a href=""><img src="" alt="" /></a>
+        </div>
+        <div class="today">
+          <a href=""><img src="" alt="" /></a>
+        </div>
+      </div>
+    </div>
+  </body>
 </html>
