+++ src/main/resources/templates/login_login.html
@@ -57,7 +57,7 @@
 		<!-- 로그인 시작하는 부분 -->
 		<div class="loginbox">
 			<div class="loginbox_form_button">
-				<form method="post" action="/login/login">
+				<form method="post" action="/member/login">
 					<fieldset>
 						<h2>LOGIN</h2>
 						<!-- ID -->
