--- src/main/resources/templates/login_socialjoin.html
@@ -1,188 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/resources/css/nav.css" />
-    <link rel="stylesheet" href="/resources/css/globalfont.css" />
-    <link rel="stylesheet" href="/resources/css/login_join.css" />
-    <link rel="shortcut icon" href="/resources/img/realfavicon.ico" />
-    <title>4Tunes</title>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script>
-      $(document).ready(function () {
-        const searchKeyword = $("#search-keyword");
-        const kwSelect = $("#kw");
-        const disabledInput = $("#disabled");
-
-        searchKeyword.on("input", function () {
-          const inputValue = searchKeyword.val().toLowerCase();
-          //사용자가 영어 소문자를 칠 수 있을 거라고 가정
-          kwSelect.find("option").each(function () {
-            const optionValue = $(this).val().toLowerCase();
-            const optionText = $(this).text().toLowerCase();
-            //value값이 같은 게 포함될때, text값이 같은 게 포함될 때 두 경우 모두 보여줌.
-            if (
-              optionValue.includes(inputValue) ||
-              optionText.includes(inputValue)
-            ) {
-              $(this).show();
-            } else {
-              $(this).hide();
-            }
-          });
-
-          disabledInput
-            .val()
-            .split(",")
-            .map(function (item) {
-              const trimmedItem = item.trim();
-              kwSelect
-                .find("option")
-                .filter(function () {
-                  return (
-                    $(this).text().toLowerCase() === trimmedItem.toLowerCase()
-                  );
-                })
-                .hide();
-            });
-        });
-
-        //옵션 더블 클릭 시 #disabled에 표시
-        kwSelect.on("dblclick", function () {
-          const selectedOptions = kwSelect
-            .find("option:selected")
-            .map(function () {
-              return $(this).text();
-            })
-            .get();
-
-          disabledInput.val(function (index, value) {
-            return value
-              ? value + "/ " + selectedOptions.join("/ ")
-              : selectedOptions.join("/ ");
-          });
-          kwSelect.find("option:selected").hide();
-        });
-      });
-    </script>
-  </head>
-
-  <body>
-  
- <script>
-  window.addEventListener('DOMContentLoaded', function() {
-    // URL에서 쿼리 파라미터를 가져옴
-    const urlParams = new URLSearchParams(window.location.search);
-    
-    // 이메일과 이름을 가져옴
-    const email = urlParams.get('email');
-    const name = urlParams.get('name');
-    
-    // 이메일과 이름 필드에 값을 할당하고 읽기 전용으로 설정
-    const emailField = document.getElementById('join-email');
-    const nameField = document.getElementById('join-name');
-    
-    emailField.value = email;
-    nameField.value = name;
-    
-    emailField.setAttribute('readonly', 'readonly');
-    nameField.setAttribute('readonly', 'readonly');
-  });
-</script>
-  
-  
-    <div id="nav">
-      <div class="banner">
-        <a href="/"><img src="/resources/img/banner.png" /></a>
-      </div>
-      <div id="login">
-        <div>
-          <a href="/nav/login"
-            ><img class="login_img" src="/resources/img/usertest.png" />LOGIN</a
-          >
-        </div>
-        <br />
-        <div>
-          <a href="/nav/mypage"
-            ><img class="mypage_img" src="/resources/img/file.png" />MYPAGE</a
-          >
-        </div>
-      </div>
-      <div class="search-container">
-        <span class="search-icon"
-          ><img src="/resources/img/search.png" class="search-icon"
-        /></span>
-        <input type="text" class="search-input" placeholder="키워드 입력" />
-      </div>
-      <div><a href="/nav/suggested">맞춤추천</a></div>
-      <div><a href="/nav/chart">인기차트</a></div>
-      <div><a href="/nav/community">커뮤니티</a></div>
-      <div><a href="/nav/membership">멤버십</a></div>
-    </div>
-
-    <div class="main">
-      <!-- 로그인 시작하는 부분 -->
-      <div class="loginbox">
-        <form action="/login/insertuser">
-          <fieldset>
-            <h2>JOIN</h2>
-            <div class="join">
-              <span class="join-title">ID</span>
-                    <input type="text" id="join-email" placeholder="Email를 입력해 주세요" value="${email}" readonly />
-              />
-            </div>
-            <div class="join">
-              <span class="join-title">PW</span>
-              <input
-                type="password"
-                id="join-pw"
-                placeholder="비밀번호를 입력해 주세요"
-              />
-            </div>
-            <div class="join">
-              <span class="join-title">이름</span>
-              <input
-                type="text"
-                id="join-name"
-                placeholder="이름을 입력해 주세요"
-              />
-            </div>
-            <div class="join-keyword">
-              <p class="join-title">KeyWord</p>
-            </div>
-            <div class="join-kw">
-              <input
-                type="search"
-                id="search-keyword"
-                placeholder="내용을 입력해주세요."
-              />
-            </div>
-            <div class="join-kw">
-             	 <select name="keyword" id="kw" multiple>
- 				 <option th:each="keyword : ${keywordlist}" th:value="${keyword}" th:text="${keyword}"></option>
-				</select>
-            </div>
-            <div class="join-kw" id="selected_keyword">
-              <input
-                type="text"
-                id="disabled"
-                placeholder="선택된 내용 보여주기"
-                disabled
-              />
-            </div>
-          </fieldset>
-          <div class="join-btn">
-            <div id="find_join">
-              <button type="submit">Join</button>
-            </div>
-            <div id="find_cancel">
-              <button type="submit">Cancel</button>
-            </div>
-          </div>
-        </form>
-      </div>
-    </div>
-  </body>
-</html>
